service/requirements.txt
cffi>=1.14.5
dasbus>=1.6
psutil>=5.8.0
docstring_parser>=0.13
sphinx>=4.5.0
sphinx_rtd_theme>=1.0.0
sphinxemoji>=0.2.0
jsonschema>=3.2.0
pyyaml>=6.0.0
setuptools>=53.0.0
pygments>=2.13.0
sphinx-tabs>=3.3.1
tap.py[yaml]>=3.1
